+++ apps/ranger-next/components/AppLayout/appLayout.tsx
@@ -1,0 +1,12 @@
+import Header from '@/components/Header'
+
+const AppLayout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <div>
+      <Header />
+      {children}
+    </div>
+  )
+}
+
+export default AppLayout