+++ apps/ranger-next/pages/_app.tsx
@@ -2,7 +2,7 @@ import { withApollo } from '@ranger-theme/core'
 import type { AppProps } from 'next/app'
 import Head from 'next/head'
 
-import Header from '@/components/Header'
+import AppLayout from '@/components/AppLayout'
 
 const App = ({ Component, pageProps }: AppProps) => {
   console.info('app is bootstrap...')
@@ -15,10 +15,9 @@ const App = ({ Component, pageProps }: AppProps) => {
           content="width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=1, shrink-to-fit=no, viewport-fit=cover"
         />
       </Head>
-      <div>
-        <Header />
+      <AppLayout>
         <Component {...pageProps} />
-      </div>
+      </AppLayout>
     </>
   )
 }