+++ apps/ranger-next/pages/checkout.tsx
@@ -1,0 +1,9 @@
+const Checkout = () => {
+  return (
+    <div className="checkout">
+      <p>Checkout</p>
+    </div>
+  )
+}
+
+export default Checkout