+++ apps/ranger-next/pages/index.tsx
@@ -1,14 +1,9 @@
-import { useEffect } from 'react'
 import { fetchEdege, HeadElement, HtmlELement, ScriptElement } from '@ranger-theme/adobe-edege'
 
 const url: string = process.env.NEXT_PUBLIC_EDEGE_URL
 const host: string = process.env.NEXT_PUBLIC_HOST_URL
 
 const Home = ({ html }: { html: string }) => {
-  useEffect(() => {
-    window?.edegeLoadPage?.()
-  }, [])
-
   return (
     <div>
       <HeadElement host={host} html={html} url={url} />