+++ packages/ranger-adobe-edege/src/ScriptElement/scriptElement.tsx
@@ -1,8 +1,7 @@
-import type { FC } from 'react'
-import { useEffect, useState } from 'react'
+import React, { useEffect, useState } from 'react'
+import { Helmet } from 'react-helmet'
 import type { HTMLReactParserOptions } from 'html-react-parser'
 import parse from 'html-react-parser'
-import Head from 'next/head'
 
 export interface ScriptElementProps {
   html: string
@@ -15,6 +14,7 @@ const options: HTMLReactParserOptions = {
   replace: (node: any) => {
     if (node.type === 'script') {
       const elType: string = node?.attribs?.type ?? ''
+
       if (!['module', 'javascript'].includes(elType)) {
         return <></>
       }
@@ -22,19 +22,28 @@ const options: HTMLReactParserOptions = {
   }
 }
 
-const ScriptElement: FC<ScriptElementProps> = ({ html, url }) => {
+const ScriptElement: React.FC<ScriptElementProps> = ({ html, url }) => {
   const [isRender, setIsRender] = useState<boolean>(false)
   const head = html.match(/<head[^>]*>([\s\S]*?)<\/head>/g)
   const title = html.match(/<title[^>]*>([\s\S]*?)<\/title>/g)
   const scriptEle = (head?.[0] ?? '').match(/<script[^>]*>([\s\S]*?)<\/script>/g) || []
   const titleEle = title?.[0] ?? ''
+  const components = parse(`${scriptEle.join('')}${titleEle}`, options) as any[]
 
   useEffect(() => {
     window.edegeURL = url
     setIsRender(true)
   }, [])
 
-  return <Head>{isRender && <>{parse(`${scriptEle.join('')}${titleEle}`, options)}</>}</Head>
+  return isRender ? (
+    <Helmet>
+      {components.map((component: any) => {
+        if (!React.isValidElement(component)) return null
+        if (component?.type === React.Fragment) return null
+        return component
+      })}
+    </Helmet>
+  ) : null
 }
 
 export default ScriptElement